Move `JsonPath.ts` file to new `jsonPath` directory
